Migrated code to new ClientRoleService
