+++ src/main/kotlin/io/craigmiller160/authmanagementservice/service/ClientService.kt
@@ -27,8 +27,6 @@ class ClientService (
 
     private val encoder = BCryptPasswordEncoder()
 
-    // TODO validate inputs
-
     fun generateGuid(): String {
         return UUID.randomUUID().toString()
     }
@@ -83,24 +81,5 @@ class ClientService (
                 ?: throw EntityNotFoundException("Role not found for ClientID $clientId and RoleID $roleId")
         val finalRole = role.copy(id = roleId, clientId = clientId)
         return roleRepo.save(finalRole)
-    }
-
-    fun createRole(clientId: Long, role: Role): Role {
-        val finalRole = role.copy(id = 0, clientId = clientId)
-        return roleRepo.save(finalRole)
-    }
-
-    @Transactional
-    fun deleteRole(clientId: Long, roleId: Long): Role {
-        val existing = roleRepo.findByClientIdAndId(clientId, roleId)
-                ?: throw EntityNotFoundException("Role not found for ClientID $clientId and RoleID $roleId")
-        clientUserRoleRepo.deleteAllByRoleIdAndClientId(roleId, clientId)
-        roleRepo.deleteByClientIdAndId(clientId, roleId)
-        return existing
-    }
-
-    fun getRoles(clientId: Long): RoleList {
-        val roles = roleRepo.findAllByClientIdOrderByName(clientId)
-        return RoleList(roles)
     }
 }
